Revive concatenation for arbitrary duration recordings
This is required to concatenate the files from the TNC (PNG/Indonesian)
recordings. Is working except that the spectral indices are not filling
in missing recordings. This is work yet to be done.

+++ AudioAnalysis/AudioAnalysisTools/Indices/SummaryIndexValues.cs
@@ -115,11 +115,11 @@ namespace AudioAnalysisTools.Indices
         public double EntropyOfCoVSpectrum { get; set; }
 
         // meaningless when calculated over short
-        public int ClusterCount { get; set; }
+        public double ClusterCount { get; set; }
 
         // public TimeSpan AvgClusterDuration { get; set; }
 
-        public int ThreeGramCount { get; set; }
+        public double ThreeGramCount { get; set; }
 
         // public double SptPerSecond { get; set; }
